refactor: lift traced circuit helpers out of surface
