+++ python/quantum-pecos/src/pecos/qec/dem.py
@@ -43,6 +43,10 @@ from typing import TYPE_CHECKING, Any
 
 from pecos_rslib.qec import DetectorErrorModel as _RustDetectorErrorModel
 
+from pecos._traced_circuit import (
+    measurement_ids_in_execution_order,
+    normalize_traced_tick_circuit,
+)
 from pecos.qec.dem_spec import GuppyDemBuild, ResultRef, _resolve_dem_specs
 
 if TYPE_CHECKING:
@@ -368,7 +372,6 @@ class _DetectorErrorModelMixin:
             scalar ``result(tag, measure(q))`` in straight-line programs; the
             runtime-loop case (per-occurrence binding) remains deferred.
         """
-        from pecos.qec.surface.circuit_builder import normalize_traced_qis_tick_circuit
         from pecos.tracing import trace_program_to_tick_circuit
 
         # Tag-referenced detectors require the compiled HUGR (to recover the
@@ -430,7 +433,7 @@ class _DetectorErrorModelMixin:
         # analysis: normalize parameterized Clifford rotations to named gates,
         # stamp stable MeasIds onto measurement gates, and fail loudly if raw
         # traced-QIS rotations survived normalization.
-        normalize_traced_qis_tick_circuit(tc, context="DetectorErrorModel.from_guppy")
+        normalize_traced_tick_circuit(tc, context="DetectorErrorModel.from_guppy")
 
         # Resolve `result_tags` -> record offsets via Rust (sound HUGR
         # extraction + runtime-loop guard via static-vs-traced measurement
@@ -438,8 +441,6 @@ class _DetectorErrorModelMixin:
         # contain `result_tags`; the downstream Rust DEM builder is unchanged.
         if needs_tags:
             from pecos_rslib import resolve_result_tags_for_guppy
-
-            from pecos.qec.surface.decode import _extract_measurement_meas_ids
 
             source_ids_json = tc.get_meta("qis_source_measurement_ids") or tc.get_meta("guppy_source_measurement_ids")
             source_measurement_ids = json.loads(source_ids_json) if source_ids_json else []
@@ -449,7 +450,7 @@ class _DetectorErrorModelMixin:
                 observables_json,
                 hugr_bytes,
                 source_measurement_ids,
-                _extract_measurement_meas_ids(tc),
+                measurement_ids_in_execution_order(tc),
             )
 
         # Hand the caller's metadata to the Rust builder verbatim; it owns all
@@ -509,9 +510,7 @@ def _validated_source_measurement_ids(circuit: Any) -> list[int]:
         msg = "source measurement identities must be a list of non-negative integers"
         raise ValueError(msg)
 
-    from pecos.qec.surface.decode import _extract_measurement_meas_ids
-
-    runtime_measurement_ids = _extract_measurement_meas_ids(circuit)
+    runtime_measurement_ids = measurement_ids_in_execution_order(circuit)
     if (
         len(source_measurement_ids) != len(set(source_measurement_ids))
         or len(runtime_measurement_ids) != len(set(runtime_measurement_ids))
@@ -811,7 +810,6 @@ def build_dem_from_guppy(
     Measurement-dependent quantum control remains unsupported because one
     captured execution is not a static circuit model.
     """
-    from pecos.qec.surface.circuit_builder import normalize_traced_qis_tick_circuit
     from pecos.tracing import _trace_program_to_tick_circuit_with_result_traces
 
     referenced_tags = sorted(
@@ -836,7 +834,7 @@ def build_dem_from_guppy(
         max_hosted_tick_separation=max_hosted_tick_separation,
         allow_raw_measurement_id_fallback=False,
     )
-    normalize_traced_qis_tick_circuit(circuit, context="build_dem_from_guppy")
+    normalize_traced_tick_circuit(circuit, context="build_dem_from_guppy")
     result_traces = _compiler_certified_result_traces(
         generator_layout,
         hugr_bytes,
