+++ python/quantum-pecos/src/pecos/qec/surface/circuit_builder.py
@@ -3001,109 +3001,6 @@ def generate_tick_circuit_from_patch(
     return renderer.render(ops, allocation, patch, num_rounds, basis)
 
 
-def normalize_traced_qis_tick_circuit(
-    tick_circuit: object,
-    *,
-    context: str = "traced-QIS DEM construction",
-    simplify_single_qubit_clifford_chains: bool = True,
-) -> None:
-    """Normalize a traced-QIS TickCircuit before DEM/DAG analysis.
-
-    Selene/QIS traces may contain parameterized Clifford rotations such as
-    ``RZZ(pi/2)``. Fault analysis and replacement-branch noise models operate
-    on the named Clifford gates (``SZZ`` / ``SZZdg``), so callers should run
-    this helper at every traced-QIS boundary before converting to a DAG.
-    """
-    _call_required_tick_circuit_method(tick_circuit, "lower_clifford_rotations", context)
-    if simplify_single_qubit_clifford_chains:
-        _call_required_tick_circuit_method(
-            tick_circuit,
-            "simplify_single_qubit_clifford_chains",
-            context,
-        )
-    _call_required_tick_circuit_method(tick_circuit, "assign_missing_meas_ids", context)
-    assert_traced_qis_tick_circuit_dem_ready(tick_circuit, context=context)
-
-
-def assert_traced_qis_tick_circuit_dem_ready(
-    tick_circuit: object,
-    *,
-    context: str = "traced-QIS DEM construction",
-) -> None:
-    """Fail loudly if raw traced-QIS rotations survived normalization."""
-    offenders = _raw_traced_qis_rzz_gates(tick_circuit, context=context)
-    if not offenders:
-        return
-
-    preview = "; ".join(offenders[:5])
-    suffix = f"; ... {len(offenders) - 5} more" if len(offenders) > 5 else ""
-    msg = (
-        f"{context}: traced-QIS circuit still contains raw RZZ gates after Clifford "
-        "normalization. DEM/DAG analysis expects Clifford RZZ(pi/2) and "
-        "RZZ(-pi/2) gates to be lowered to SZZ/SZZdg before noise attachment "
-        "and fault propagation. Call normalize_traced_qis_tick_circuit(...) "
-        "before to_dag_circuit(), or extend lower_clifford_rotations() for the "
-        f"runtime-emitted angle. First offending gates: {preview}{suffix}"
-    )
-    raise ValueError(msg)
-
-
-def _call_required_tick_circuit_method(tick_circuit: object, method_name: str, context: str) -> None:
-    method = getattr(tick_circuit, method_name, None)
-    if not callable(method):
-        msg = f"{context}: expected a TickCircuit with callable {method_name}()."
-        raise TypeError(msg)
-    method()
-
-
-def _raw_traced_qis_rzz_gates(tick_circuit: object, *, context: str) -> list[str]:
-    try:
-        num_ticks = int(tick_circuit.num_ticks())  # type: ignore[attr-defined]
-    except AttributeError as exc:
-        msg = f"{context}: expected a TickCircuit with num_ticks() before DEM/DAG analysis."
-        raise TypeError(msg) from exc
-
-    offenders: list[str] = []
-    for tick_index in range(num_ticks):
-        try:
-            tick = tick_circuit.get_tick(tick_index)  # type: ignore[attr-defined]
-        except AttributeError as exc:
-            msg = f"{context}: expected a TickCircuit with get_tick() before DEM/DAG analysis."
-            raise TypeError(msg) from exc
-        try:
-            gate_batches = tick.gate_batches()
-        except AttributeError as exc:
-            msg = f"{context}: expected TickCircuit ticks with gate_batches() before DEM/DAG analysis."
-            raise TypeError(msg) from exc
-        for gate_index, gate in enumerate(gate_batches):
-            if _gate_type_name(gate) != "RZZ":
-                continue
-            qubits = [int(q) for q in getattr(gate, "qubits", [])]
-            offenders.append(
-                f"tick={tick_index} gate={gate_index} qubits={qubits} angles={_gate_angles_for_message(gate)}",
-            )
-    return offenders
-
-
-def _gate_type_name(gate: object) -> str:
-    gate_type = getattr(gate, "gate_type", "")
-    return str(getattr(gate_type, "name", str(gate_type).rsplit(".", maxsplit=1)[-1]))
-
-
-def _format_gate_angle(angle: object) -> str:
-    try:
-        return repr(float(angle))
-    except (TypeError, ValueError):
-        return repr(angle)
-
-
-def _gate_angles_for_message(gate: object) -> list[str]:
-    angles = getattr(gate, "angles", None)
-    if angles is None:
-        angles = getattr(gate, "params", [])
-    return [_format_gate_angle(angle) for angle in angles]
-
-
 def get_detector_descriptors_from_tick_circuit(
     tick_circuit: TickCircuit,
     patch: SurfacePatch,
