+++ python/quantum-pecos/tests/qec/test_traced_qis_clifford_pipeline.py
@@ -7,8 +7,8 @@ import math
 import random
 
 import pytest
+from pecos._traced_circuit import normalize_traced_tick_circuit
 from pecos.qec.surface import SurfacePatch
-from pecos.qec.surface.circuit_builder import normalize_traced_qis_tick_circuit
 from pecos.qec.surface.decode import _build_surface_tick_circuit_for_native_model
 from pecos.quantum import TickCircuit
 from pecos_rslib_exp import (
@@ -82,7 +82,7 @@ TICK_2Q_METHODS = {
 def build_lowered_traced_qis_surface_code(rounds=3):
     patch = SurfacePatch.create(distance=3)
     tc = _build_surface_tick_circuit_for_native_model(patch, rounds, "Z", circuit_source="traced_qis")
-    normalize_traced_qis_tick_circuit(tc, context="test traced-QIS surface code")
+    normalize_traced_tick_circuit(tc, context="test traced-QIS surface code")
     return tc
 
 
@@ -216,11 +216,11 @@ def test_lowered_traced_qis_pipeline_sampling_and_catalog_smoke():
     assert len(first_fault.faults) == 1
 
 
-def test_normalize_traced_qis_tick_circuit_lowers_clifford_rzz():
+def test_normalize_traced_tick_circuit_lowers_clifford_rzz():
     tc = TickCircuit()
     tc.tick().rzz(math.pi / 2, [(0, 1)])
 
-    normalize_traced_qis_tick_circuit(tc, context="test Clifford RZZ normalization")
+    normalize_traced_tick_circuit(tc, context="test Clifford RZZ normalization")
 
     gate_names = [
         gate.gate_type.name for tick_index in range(tc.num_ticks()) for gate in tc.get_tick(tick_index).gate_batches()
@@ -229,12 +229,12 @@ def test_normalize_traced_qis_tick_circuit_lowers_clifford_rzz():
     assert "SZZ" in gate_names
 
 
-def test_normalize_traced_qis_tick_circuit_rejects_raw_rzz_after_lowering():
+def test_normalize_traced_tick_circuit_rejects_raw_rzz_after_lowering():
     tc = TickCircuit()
     tc.tick().rzz(math.pi / 4, [(0, 1)])
 
     with pytest.raises(ValueError, match="still contains raw RZZ"):
-        normalize_traced_qis_tick_circuit(tc, context="test non-Clifford RZZ normalization")
+        normalize_traced_tick_circuit(tc, context="test non-Clifford RZZ normalization")
 
 
 def _gate_names(tc):
@@ -292,22 +292,22 @@ def test_tick_circuit_pass_bindings_simplify_single_qubit_clifford_chains():
     assert _gate_names(tc) == ["F"]
 
 
-def test_normalize_traced_qis_tick_circuit_simplifies_single_qubit_clifford_chains():
+def test_normalize_traced_tick_circuit_simplifies_single_qubit_clifford_chains():
     tc = TickCircuit()
     tc.tick().sx([0])
     tc.tick().sz([0])
 
-    normalize_traced_qis_tick_circuit(tc, context="test one-qubit Clifford normalization")
+    normalize_traced_tick_circuit(tc, context="test one-qubit Clifford normalization")
 
     assert _gate_names(tc) == ["F"]
 
 
-def test_normalize_traced_qis_tick_circuit_can_skip_single_qubit_clifford_simplification():
+def test_normalize_traced_tick_circuit_can_skip_single_qubit_clifford_simplification():
     tc = TickCircuit()
     tc.tick().sx([0])
     tc.tick().sz([0])
 
-    normalize_traced_qis_tick_circuit(
+    normalize_traced_tick_circuit(
         tc,
         context="test one-qubit Clifford normalization",
         simplify_single_qubit_clifford_chains=False,
